Introduces CompleteOfcHand interface (12% speedup).
According to the profiler, we were spending a lot of time checking and
re-checking that hands were complete. By splitting the "complete hands only"
methods of OfcHand into a subinterface, we can perform the check only once at
runtime -- and the remaining checks can be done by the compiler's type-checker.

This also helps avoid the "methods that only work sometimes" anti-pattern.